+++ src/reporter.ts
@@ -92,6 +92,8 @@ export class RPReporter implements Reporter {
 
   constructor(config: ReportPortalConfig) {
     this.config = {
+      uploadTrace: true,
+      uploadVideo: true,
       ...config,
       launchId: process.env.RP_LAUNCH_ID || config.launchId,
     };
@@ -464,7 +466,11 @@ export class RPReporter implements Reporter {
 
     // TODO: cover with tests
     if (result.attachments?.length) {
-      const attachmentsFiles = await getAttachments(result.attachments);
+      const { uploadVideo, uploadTrace } = this.config;
+      const attachmentsFiles = await getAttachments(result.attachments, {
+        uploadVideo,
+        uploadTrace,
+      });
 
       attachmentsFiles.map((file) => {
         this.sendLog(testItemId, {