+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Added
 - `launchId` option to the config to attach run results to an existing launch. Related to parallel execution on one and several machines [#86](https://github.com/reportportal/agent-js-playwright/issues/86).
+- `uploadVideo` and `uploadTrace` options to the config to disable auto-attaching the Playwright's [video](https://playwright.dev/docs/api/class-testoptions#test-options-video) and [trace](https://playwright.dev/docs/api/class-testoptions#test-options-trace). Addressed [#98](https://github.com/reportportal/agent-js-playwright/issues/98).
 
 ## [5.0.11] - 2023-04-13
 ### Fixed